% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogRank.R
\name{binarize}
\alias{binarize}
\title{binarize a vector by the mediate value}
\usage{
binarize(x = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
The binarized vector
}
\description{
binarize a vector by the mediate value
}
