% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerPrognosis_LncRNAData.R
\name{CancerPrognosis_LncRNAData}
\alias{CancerPrognosis_LncRNAData}
\title{Evaluate Cancer Prognosis of Long non-coding RNA expression data to compute the risk scores}
\usage{
CancerPrognosis_LncRNAData(data, methods)
}
\arguments{
\item{data}{data to be computed for cancer Prognosis risk scores; either a data matrix or ExpressionSet object. 
If it is a data matrix, rows= lncRNAs annotated with gene gene symbols and columns=tems/samples .}

\item{methods}{A set of methods to be performed in the 6 methods:"HOTAIR", "MALAT1", "DSCAM-AS1", "lncRNA5","lncRNA6","lncRNA12".}
}
\value{
A dataframe object with rows for samples and columns which represent dataset used and its corresponding methods
}
\description{
Cancer Prognosis to compute the risk scores for long non-coding RNA data based on the 6 methods: "HOTAIR", "MALAT1", "DSCAM-AS1",
           "lncRNA5","lncRNA6","lncRNA12".
}
\examples{
data(TCGA500)
methods <- c("HOTAIR", "MALAT1", "DSCAM-AS1", "lncRNA12","lncRNA6","lncRNA5")
res = CancerPrognosis_LncRNAData(data=TCGA500, methods=methods)
}
\references{
\itemize{
 \item \strong{HOTAIR}: Paw lowska E, Szczepanska J, Blasiak J. The Long Noncoding RNA HOTAIR in Breast Cancer: Does Autophagy Play a Role? International journal of molecular sciences. 2017;18(11):2317.
 \item \strong{MALAT1}: Wang Z, Katsaros D, Biglia N, Shen Y, Fu Y, Loo LW, et al. High expression of long non-coding RNA MALAT1 in breast cancer is associated with poor relapse-free survival. Breast cancer research and treatment. 2018; p. 1-11.
 \item \strong{DSCAM-AS1}: Niknafs YS, Han S, Ma T, Speers C, Zhang C, Wilder-Romans K, et al. The lncRNA landscape of breast cancer reveals a role for DSCAM-AS1 in breast cancer progression. Nature communications. 2016;7:12791.
 \item \strong{lncRNA12}: Zhou M, Zhong L, Xu W, Sun Y, Zhang Z, Zhao H, et al. Discovery of potential prognostic long non-coding RNA biomarkers for predicting the risk of tumor recurrence of breast cancer patients. Scientific reports. 2016;6:31038.
 \item \strong{lncRNA6}: Zhong L, Lou G, Zhou X, Qin Y, Liu L, Jiang W. A six-long non-coding RNAs signature as a potential prognostic marker for survival prediction of ER-positive breast cancer patients. Oncotarget. 2017;8(40):67861.
 \item \strong{lncRNA5}: Li J, Wang W, Xia P, Wan L, Zhang L, Yu L, et al. Identification of a five-lncRNA signature for predicting the risk of tumor recurrence in breast cancer patients. International journal of cancer. 2018;.
 }
}
