% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerPrognosis_RNAData.R
\name{CancerPrognosis_RNAData}
\alias{CancerPrognosis_RNAData}
\title{Evaluate Cancer Prognosis of coding RNA expression data to compute the risk scores in 13 methods}
\usage{
CancerPrognosis_RNAData(data, platform = "custom", methods, RNASig = NULL)
}
\arguments{
\item{data}{data to be computed for Cancer Prognosis risk scores; either a data matrix or ExpressionSet object. If it is a data matrix, rows= probes/genes annotated with Entrez ID and columns=terms/samples .}

\item{platform}{The technical platform for data, "affy" or "agilent" or "custom". Defualt "custom" represnts unknow}

\item{methods}{A set of methods to be performed in the 13 methods: "AURKA", "ESR1", "ERBB2", "GGI", "GENIUS", "Endopredict", "OncotypeDx","TAMR13", "PIK3CAGS", "GENE70", "rorS", "RNAmodel","Ensemble",
The "Ensemble" method is the average of the five methods, i.e. "GENIUS", "Endopredict", "OncotypeDx", "GENE70", "rorS".}

\item{RNASig}{A dataframe represnts the customized signatures provided for the "RNAmodel" method. The columns should include "Gene.symbol","EntrezGene.ID", "weight". Default NULL, the default 30 RNA signatures will be used.}
}
\value{
A dataframe object with rows for samples and columns which represent dataset used and its corresponding methods.
}
\description{
Cancer Prognosis to compute the risk scores for coding RNA expression data based on existing 13 methods: 
           "AURKA", "ESR1", "ERBB2", "GGI", "GENIUS", "Endopredict", "OncotypeDx","TAMR13", "PIK3CAGS", "GENE70", "rorS", "RNAmodel","Ensemble"
}
\examples{
\dontrun{
library("breastCancerMAINZ")
data("mainz")
methods <- c("AURKA", "ESR1", "ERBB2", "GGI", "GENIUS", "Endopredict", "OncotypeDx", 
            "TAMR13", "PIK3CAGS", "GENE70", "rorS", "RNAmodel", "Ensemble")
res = CancerPrognosis_RNAData(data=mainz, platform="custom", methods=methods)
}

}
