% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerSubtypes.R
\name{CancerSubtypes}
\alias{CancerSubtypes}
\title{Evaluate Cancer CancerSubtyping methods of mRNA, miRNA or multiomics data based on the running time, the average Silhouette score, and the p-value of Logrank test}
\usage{
CancerSubtypes(
  dn = NULL,
  omics = NULL,
  methods = NULL,
  fileFolder = NULL,
  logFile = NULL
)
}
\arguments{
\item{dn}{datasets to be computed for cancer subtypes; a string vector.}

\item{omics}{The types of data, "mRNA" or "miRNA" or "multiomics".}

\item{methods}{A set of methods to be performed in the 11 methods: "PAM50","IntClust","CC","CNMF","iCluster","SNF","SNF-CC","WSNF","CIMLR","PINS","NEMO".}

\item{fileFolder}{A file folder name provided for saving results.}

\item{logFile}{A file name provided for saving log information.}
}
\value{
a list contains timeTable matrix, silTable matrix, and pvalueTable matrix
}
\description{
CancerSubtypes to compute the running time, the average Silhouette score, and the p-value of Logrank test for mRNA, miRNA or multiomics data based on existing 11 methods: 
           "PAM50","IntClust","CC","CNMF","iCluster","SNF","SNF-CC","WSNF","CIMLR","PINS","NEMO"
}
\examples{
\dontrun{
dn = c("TCGA","UK","HEL","GSE19783")
methods = c("PAM50","IntClust","CC","CNMF","iCluster","SNF","SNF-CC","WSNF","CIMLR","PINS","NEMO")
omics = "mRNA"
res = CancerSubtypes(dn, omics, methods, fileFolder=omics, logFile = omics)
}

}
