% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cindex.R
\name{Cindex}
\alias{Cindex}
\title{C-index calculation}
\usage{
Cindex(data, survival, PValue = FALSE, outputFolder = NULL)
}
\arguments{
\item{data}{A dataframe object with rows for samples and columns represent corresponding methods. A return value from CancerPrognosis_xxx() function}

\item{survival}{A dataframe object which contains variables (columns) representing for survival time and event. The rows are the samples.}

\item{PValue}{if ture, return the object of function concordance.index().}

\item{outputFolder}{(Optional) A desired folder to put the results or "output" by default}
}
\value{
This function is used for its side-effect.
A plot for overall Concordance Index. For C-Index for each method, please refer in the outputFolder
}
\description{
Calculating Concordance Indices for the evaluation results of Cancer Prognosis methods
}
\examples{
library("breastCancerMAINZ")
data("mainz")
methods <- c("AURKA", "ESR1", "ERBB2", "GGI", "GENIUS", "Endopredict", "OncotypeDx", 
            "TAMR13", "PIK3CAGS", "GENE70", "rorS", "RNAmodel", "Ensemble")
sampleInfo= pData(mainz)           
survival=data.frame(time=sampleInfo$t.dmfs,event=sampleInfo$e.dmfs, row.names=sampleInfo$samplename)
\dontrun{
res = CancerPrognosis_RNAData(data=mainz, platform="custom", methods=methods)
CIs = Cindex(data=res, survival,outputFolder="./mainz")
}
}
