% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringMethod.R
\name{ExecuteCIMLR}
\alias{ExecuteCIMLR}
\title{Execute CIMLR (Cancer Integration via Multikernel Learning)}
\usage{
ExecuteCIMLR(datasets, clusterNum, k = 10, ncore = 0, plot = TRUE)
}
\arguments{
\item{datasets}{A data matrix or a list containing data matrices. For each data matrix, the rows represent genomic features, and the columns represent samples.}

\item{clusterNum}{Number of subtypes for the samples}

\item{k}{tuning parameter in CIMLR. A default of 10 is performed.}

\item{ncore}{ratio of the number of cores to be used when computing the multi-kernel}

\item{plot}{Logical value. If true, draw the heatmap for the distance matrix with samples ordered to form clusters.}
}
\value{
A list with the following elements.
\itemize{
 \item \strong{group} : A vector represent the group of cancer subtypes. The order is corresponding to the the samples in the data matrix.\cr
  
  This is the most important result for all clustering methods, so we place it as the first component. The format of group 
  is consistent across different algorithms and therefore makes it convenient for downstream analyses. Moreover, the format
  of group is also compatible with the K-means result and the hclust (after using the cutree() function).
  
 \item \strong{distanceMatrix} : It is a sample similarity matrix. The more large value between samples in the matrix, the more similarity the samples are.
  
  We extracted this matrix from the algorithmic procedure because it is useful for similarity analysis among the samples based on the clustering results.
  
 \item \strong{timing} : The running time.
 }
}
\description{
CIMLR calculates the similarity between patients in multi-omic data by combining a set of Gaussian kernels for each single-omic data.
}
\details{
The R package "CIMLR" should be installed.


 If the data is a list containing the matched mutli-genomics data matrices like the input as "ExecuteiCluster()" and "ExecuteSNF()",
  The data matrices in the list are concatenated according to samples. The concatenated data matrix is the samples with a long features (all features in the data list).
  Our purpose is to make convenient comparing the different method with same dataset format. See examples.
}
\examples{
data(GeneExp)
result=ExecuteCIMLR(GeneExp,clusterNum=5)
result$group

}
\references{
Daniele Ramazzotti, Avantika Lal, Bo Wang, Serafim Batzoglou, and Arend Sidow. Multi-omic tumor data reveal diversity of molecular mechanisms that correlate with survival. Nature communications, 9(1):4453, 2018.
}
\seealso{
\code{\link{CIMLR}}
}
