% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringMethod.R
\name{ExecuteNEMO}
\alias{ExecuteNEMO}
\title{Execute NEMO (NEighborhood based Multi-Omics clustering)}
\usage{
ExecuteNEMO(datasets, clusterNum, num.neighbors = 50, plot = TRUE)
}
\arguments{
\item{datasets}{A data matrix or a list containing data matrices. For each data matrix, the rows represent genomic features, and the columns represent samples.}

\item{clusterNum}{Number of subtypes for the samples}

\item{num.neighbors}{The number of neighbors to use for each omic. It can either be a number, a list of numbers
or NA. If it is a number, this is the number of neighbors used for all omics. If this is a list,
the number of neighbors are taken for each omic from that list. If it is NA, each omic chooses the
number of neighbors to be the number of samples divided by NUM.NEIGHBORS.RATIO.}

\item{plot}{Logical value. If true, draw the heatmap for the distance matrix with samples ordered to form clusters.}
}
\value{
A list with the following elements.
\itemize{
 \item \strong{group} : A vector represent the group of cancer subtypes. The order is corresponding to the the samples in the data matrix.\cr
  
  This is the most important result for all clustering methods, so we place it as the first component. The format of group 
  is consistent across different algorithms and therefore makes it convenient for downstream analyses. Moreover, the format
  of group is also compatible with the K-means result and the hclust (after using the cutree() function).
  
 \item \strong{distanceMatrix} : It is a sample similarity matrix. The more large value between samples in the matrix, the more similarity the samples are.
  
  We extracted this matrix from the algorithmic procedure because it is useful for similarity analysis among the samples based on the clustering results.
  
 \item \strong{timing} : The running time.
 }
}
\description{
NEMO is a similarity-based multi-omic clustering method based on the radial basis of function kernel and spectral clustering method. 
Different from other multi-omic clustering methods, NEMO can apply to partial data that some omics may not be measured for some patients. NEMO is simple, and faster than other multi-omics clustering methods but achieved the comparable performance to others.
}
\details{
If the data is a list containing the matched mutli-genomics data matrices like the input as "ExecuteiCluster()" and "ExecuteSNF()",
  The data matrices in the list are concatenated according to samples. The concatenated data matrix is the samples with a long features (all features in the data list).
  Our purpose is to make convenient comparing the different method with same dataset format. See examples.
}
\examples{
data(GeneExp)
result=ExecuteNEMO(GeneExp,clusterNum=3)
result$group

}
\references{
Nimrod Rappoport and Ron Shamir. Nemo: cancer subtyping by integration of partial multi-omic data. Bioinformatics, 35(18):3348-3356, 2019.
}
