% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringMethod.R
\name{ExecutePINS}
\alias{ExecutePINS}
\title{Execute PINS (Perturbation clustering for data INtegration and disease Subtyping)}
\usage{
ExecutePINS(
  datasets,
  clusterNum,
  ncore = 1,
  clusteringMethod = "kmeans",
  perturbMethod = "noise",
  iterMin = 20,
  iterMax = 200,
  madMin = 0.001,
  msdMin = 1e-06
)
}
\arguments{
\item{datasets}{A data matrix or a list containing data matrices. For each data matrix, the rows represent genomic features, and the columns represent samples.}

\item{clusterNum}{Number of subtypes for the samples}

\item{ncore}{Number of cores that the algorithm should use. Default value is 1.}

\item{clusteringMethod}{The name of built-in clustering algorithm that PerturbationClustering will use. Currently supported algorithm are kmeans, pam and hclust. Default value is "kmeans".}

\item{perturbMethod}{The clustering algorithm function that will be used instead of built-in algorithms.}

\item{iterMin}{The minimum number of iterations. Default value is 20.}

\item{iterMax}{The maximum number of iterations. Default value is 200.}

\item{madMin}{The minimum of Mean Absolute Deviation of AUC of Connectivity matrix for each k. Default value is 1e-03.}

\item{msdMin}{The minimum of Mean Square Deviation of AUC of Connectivity matrix for each k. Default value is 1e-06.}
}
\value{
A list with the following elements.
\itemize{
 \item \strong{group} : A vector represent the group of cancer subtypes. The order is corresponding to the the samples in the data matrix.\cr
  
  This is the most important result for all clustering methods, so we place it as the first component. The format of group 
  is consistent across different algorithms and therefore makes it convenient for downstream analyses. Moreover, the format
  of group is also compatible with the K-means result and the hclust (after using the cutree() function).
  
 \item \strong{timing} : The running time.
 }
}
\description{
PINS initially clusters patients based on each omic data separately using perturbation clustering and outputs the optimal number of clusters ki, original connectivity matrix Ci and the perturbed connectivity matrix Ai, where i is the index of i-th omic data.
The R package "PINSPlus" should be installed.
}
\details{
If the data is a list containing the matched mutli-genomics data matrices like the input as "ExecuteiCluster()" and "ExecuteSNF()",
  The data matrices in the list are concatenated according to samples. The concatenated data matrix is the samples with a long features (all features in the data list).
  Our purpose is to make convenient comparing the different method with same dataset format. See examples.
}
\examples{
data(GeneExp)
result=ExecutePINS(GeneExp,clusterNum=5)
result$group

}
\references{
Tin Nguyen, Rebecca Tagett, Diana Diaz, and Sorin Draghici. A novel approach for data integration and disease subtyping. Genome research, 27(12):2025-2039, 2017.
}
\seealso{
\code{\link{PerturbationClustering}}
}
