% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{getFilePath}
\alias{getFilePath}
\title{get file path based on a given folder name and a given file name}
\usage{
getFilePath(foldername = NULL, filename = NULL)
}
\arguments{
\item{foldername}{A given folder name; if NULL, will set to the current working directory.}

\item{filename}{A given file name}
}
\value{
The full path of the file
}
\description{
get file path based on a given folder name and a given file name
}
