% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valiation.R
\name{getPvalue}
\alias{getPvalue}
\title{get p-value of the Log-rank test from predicted subtypes}
\usage{
getPvalue(time, status, group)
}
\arguments{
\item{time}{survival time of samples}

\item{status}{event status of samples}

\item{group}{predicted subtypes of samples}
}
\value{
The p-value of the Log-rank test
}
\description{
get p-value of the Log-rank test from predicted subtypes
}
\examples{
\dontrun{
res = getPvalue(dclin[,1],dclin[,2],IC10_res$group)
}

}
