% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodel.R
\name{lncRNA12model}
\alias{lncRNA12model}
\title{Evaluate Cancer Prognosis based on 12 lncRNA signatures
Cancer Prognosis to compute the risk scores for lncRNA data based on 12 lncRNA signatures}
\usage{
lncRNA12model(data, annot, RNASignature = NULL, do.mapping = FALSE)
}
\arguments{
\item{data}{data to be computed for cancer Prognosis risk scores; a data matrix, rows= lncRNA annotated with lncRNA names and columns=terms/samples.}

\item{annot}{annot holds RNA names in the data.}

\item{RNASignature}{RNA signatures, if none, will import the 12 lncRNA signatures.}

\item{do.mapping}{if do.mapping equals Ture, use probe IDs instead of gene symbols.}
}
\value{
A Numeric Vector
}
\description{
Evaluate Cancer Prognosis based on 12 lncRNA signatures
Cancer Prognosis to compute the risk scores for lncRNA data based on 12 lncRNA signatures
}
\details{
The 12 lncRNA signatures are "RP1-34M23.5","RP11-202K23.1", "RP11-560G2.1", "RP4-591L5.2","RP13-104F24.2", "RP11-506D12.5","ERVH48-1","RP4-613B23.1",  "RP11-360F5.1" ,"CTD-2031P19.5", "RP11-247A12.8", "SNHG7".
}
\examples{
data(TCGA500)
data = exprs(TCGA500) 
annot = fData(TCGA500)
res =  lncRNA12model(data,annot)
}
\references{
\itemize{
 \item \strong{lncRNA12model}: Zhou M, Zhong L, Xu W, Sun Y, Zhang Z, Zhao H, Yang L, Sun J. Discovery of potential prognostic long non-coding RNA biomarkers for predicting the risk of tumor recurrence of breast cancer patients. Sci Rep. 2016;6:31038.
}
}
