% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodel.R
\name{lncRNA5model}
\alias{lncRNA5model}
\title{Evaluate Cancer Prognosis based on 5 lncRNA signatures
Cancer Prognosis to compute the risk scores for lncRNA data based on 5 lncRNA signatures}
\usage{
lncRNA5model(data, annot, RNASignature = NULL, do.mapping = FALSE)
}
\arguments{
\item{data}{data to be computed for cancer Prognosis risk scores; a data matrix, rows= lncRNA annotated with lncRNA names and columns=terms/samples.}

\item{annot}{annot holds RNA names in the data.}

\item{RNASignature}{RNA signatures, if none, will import the 5 lncRNA signatures.}

\item{do.mapping}{if do.mapping equals Ture, use probe IDs instead of gene symbols.}
}
\value{
A Numeric Vector
}
\description{
Evaluate Cancer Prognosis based on 5 lncRNA signatures
Cancer Prognosis to compute the risk scores for lncRNA data based on 5 lncRNA signatures
}
\details{
The 5 lncRNA signatures are "RP11-524D16-A.3", "HOTAIR", "AL645608.1", "TSPOAP1-AS1", "RP11-13L2.4".
}
\examples{
data(TCGA500)
data = exprs(TCGA500) 
annot = fData(TCGA500)
res =  lncRNA5model(data,annot)
}
\references{
\itemize{
 \item \strong{lncRNA5model}: Li J, Wang W, Xia P, et al. Identification of a five‐lncRNA signature for predicting the risk of tumor recurrence in patients with breast cancer. Int J Cancer. 2018; 143: 2150‐2160.
}
}
