% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lncRNA6}
\alias{lncRNA6}
\title{Dataset: lncRNA signatures}
\format{
A data frame
}
\description{
6 lncRNA signatures
}
\details{
\itemize{
 \item Rows are lncRNAs
}
}
\examples{
data(sig.lncRNA6)
}
\keyword{datasets}
