% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodel.R
\name{lncRNA6model}
\alias{lncRNA6model}
\title{Evaluate Cancer Prognosis based on 6 lncRNA signatures
Cancer Prognosis to compute the risk scores for lncRNA data based on 6 lncRNA signatures}
\usage{
lncRNA6model(data, annot, RNASignature = NULL)
}
\arguments{
\item{data}{data to be computed for cancer Prognosis risk scores; a data matrix, rows= lncRNA annotated with lncRNA names and columns=terms/samples.}

\item{annot}{annot holds RNA names in the data.}

\item{RNASignature}{RNA signatures, if none, will import the 6 lncRNA signatures.}
}
\value{
A Numeric Vector
}
\description{
Evaluate Cancer Prognosis based on 6 lncRNA signatures
Cancer Prognosis to compute the risk scores for lncRNA data based on 6 lncRNA signatures
}
\details{
The 6 lncRNA signatures are "HAGLR","STK4-AS1", "DLEU7-AS1", "LINC00957","LINC01614", "ITPR1-AS1".
}
\examples{
data(TCGA500)
data = exprs(TCGA500) 
annot = fData(TCGA500)
res =  lncRNA6model(data,annot)
}
\references{
\itemize{
 \item \strong{lncRNA6model}: Zhong L, Lou G, Zhou X, Qin Y, Liu L, Jiang W. A six-long non-coding RNAs signature as a potential prognostic marker for survival prediction of ER-positive breast cancer patients. Oncotarget. 2017;8(40):67861.
}
}
