% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{loadData}
\alias{loadData}
\title{load R data from a given directory}
\usage{
loadData(filefolder = NULL, dataSets = NULL)
}
\arguments{
\item{filefolder}{A number}

\item{dataSets}{names of datasets; a string vector.}
}
\description{
load R data from a given directory
}
\examples{
\dontrun{
dn = c("TCGA","UK","HEL","GSE19783")
loadData("data",dn)
}

}
