% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogRank.R
\name{LogRank}
\alias{LogRank}
\title{Evaluate the predicted risk scores from benchmark methods}
\usage{
LogRank(data, survival)
}
\arguments{
\item{data}{A numeric data frame or metrix indicating the predicted risk scores from benchmark methods}

\item{survival}{A data frame holds survival time and event status}
}
\value{
The p-values of the benchmark methods
}
\description{
Evaluate the predicted risk scores from benchmark methods
}
\examples{
\dontrun{
res = LogRank(data=resMatrix[[i]], survival)
}

}
\seealso{
\code{\link{survdiff}}
}
