% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRNA10}
\alias{miRNA10}
\title{Dataset: miRNA signatures}
\format{
A data frame
}
\description{
The 10 miRNAs signatures are  "hsa-miR-144", "hsa-miR-150", "hsa-miR-210", "hsa-miR-27b", "hsa-miR-30c", "hsa-miR-342"    "hsa-miR-128a", "hsa-miR-135a", "hsa-miR-767-3p", "hsa-miR-769-3p".
Note that the miRbase version of miRNAs signatures is v9_2, please check the miRbase version of your own dataset.
}
\details{
\itemize{
 \item Rows are miRNAs
}
}
\examples{
data(sig.miRNA10)
}
\keyword{datasets}
