% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodel.R
\name{miRNA10model}
\alias{miRNA10model}
\title{Evaluate Cancer Prognosis based on 10 miRNA signatures
Cancer Prognosis to compute the risk scores for miRNA data based on 10 miRNA signatures}
\usage{
miRNA10model(data, annot, RNASignature = NULL)
}
\arguments{
\item{data}{data to be computed for cancer Prognosis risk scores; a data matrix, rows= miRNAs annotated with miRNA names and columns=terms/samples.}

\item{annot}{annot holds RNA names in the data.}

\item{RNASignature}{RNA signatures, if none, will import the 10 microRNA signatures.}
}
\value{
A Numeric Vector
}
\description{
Evaluate Cancer Prognosis based on 10 miRNA signatures
Cancer Prognosis to compute the risk scores for miRNA data based on 10 miRNA signatures
}
\details{
The 10 miRNAs signatures are  "hsa-miR-144", "hsa-miR-150", "hsa-miR-210", "hsa-miR-27b", "hsa-miR-30c", "hsa-miR-342"    "hsa-miR-128a", "hsa-miR-135a", "hsa-miR-767-3p", "hsa-miR-769-3p".
Note that the miRbase version of miRNAs signatures is v9_2, please check the miRbase version of your own dataset.
}
\examples{
data(TCGA500)
data = exprs(TCGA500)
annot = fData(TCGA500)
res =  miRNA10model(data,annot)
}
\references{
\itemize{
 \item \strong{miRNA10model}: Buffa, F. M. et al. microRNA associated progression pathways and potential therapeutic targets identified by integrated mRNA and microRNA expression profiling in breast cancer. Cancer Res. 71, 5635 (2011).
}
}
