% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodel.R
\name{RNAmodel}
\alias{RNAmodel}
\title{Evaluate Cancer Prognosis based on 37 miRNA/mRNA signatures
Cancer Prognosis to compute the risk scores for miRNA/mRNA data based on 37 microRNA/mRNA signatures}
\usage{
RNAmodel(data, annot, RNASignature = NULL)
}
\arguments{
\item{data}{data to be computed for cancer Prognosis risk scores; a data matrix, rows= miRNAs annotated with miRNA names and columns=terms/samples.}

\item{annot}{annot holds miRNA/mRNA names in the data.}

\item{RNASignature}{RNA signatures, if none, will import the 37 microRNA/mRNA signatures.}
}
\value{
A Numeric Vector
}
\description{
Evaluate Cancer Prognosis based on 37 miRNA/mRNA signatures
Cancer Prognosis to compute the risk scores for miRNA/mRNA data based on 37 microRNA/mRNA signatures
}
\details{
The 7 miRNAs signatures are  "hsa-miR-103", "hsa-miR-1307", "hsa-miR-148b", "hsa-miR-328", "hsa-miR-484", "hsa-miR-874","hsa-miR-93".
The 30 mRNA signatures are "ACSL1" ,"ADAT1", "ANKRD52","BIRC6", "CPT1A", "CXCR7","DAAM1", "DIP2B", "FAM199X", "FAM91A1", 
"FRZB",  "GLA",  "GMCL1", "HRASLS",  "HSP90AA1", "MCM10", "ME1",  "NDRG1", 
"NOTCH2NL","OTUD6B","PDSS2", "PGK1",  "PIK3CA", "PTAR1", "SMG1",  "TRIM23", "TTC3",  "UBR5",  "UBXN7", "ZFC3H1".
Note that the miRbase version of miRNAs signatures is v16, please check the miRbase version of your own dataset.
}
\examples{
data(TCGA500)
data = exprs(TCGA500)
annot = fData(TCGA500)
res =  RNAmodel(data,annot)
}
\references{
\itemize{
 \item \strong{RNAmodel}: Volinia S, Croce CM. Prognostic microRNA/mRNA signature from the integrated analysis of patients with invasive breast cancer. Proc Natl Acad Sci USA. 2013;110:7413‐7417.
}
}
