% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA500}
\alias{TCGA500}
\title{Dataset: TCGA500 ExpressionSet}
\format{
A ExpressionSet
}
\description{
The TCGA breast cancer datasets is downloaded from the TCGA data portal and the datasets consist of level 3 mRNA and miRNA, lncRNA expression data from multiple platforms.
}
\details{
\itemize{
 \item ExpressionSet
}
}
\examples{
data(TCGA500)
}
\keyword{datasets}
